+++ CHANGELOG.md
@@ -8,3 +8,7 @@ Not Released Yet
 	- **U:** Undo a mistake
 	- **X:** Disable/enable extension in real time
 	- **Esc:** Remove cursor from answer input and stop typing
+	- **Enter:** Put cursor back into answer input (only if answer input is empty)
+
+## Popup
+- Setup of some settings and hotkeys